Issue #377 - use the VMUSB drivers in place rather than as copies
Use SIS 3804 scaler support from vmusb/daqconfig with #ifdefery.
