+++ main/usb/vmusb/daqconfig/C3804.h
@@ -16,24 +16,16 @@
 	     East Lansing, MI 48824-1321
 */
 
-#ifndef __CREADOUTHARDWARE_H
+/**
+ * @file C3804.h
+ * @brief header for supporting SIS 3804 scaler in MVLC an VMUSB
+ * @note when compiling for mvlcgenerate, MVLC_GENERATOR is defined.
+ */
 #include "CReadoutHardware.h"
-#endif
-
-#ifndef __STL_STRING
 #include <string>
-#ifndef __STL_STRING
-#define __STL_STRING
-#endif
-#endif
-
-
-
-#ifndef __CRT_STDINT_H
 #include <stdint.h>
-#ifndef __CRT_STDINT_H
-#define __CRT_STDINT_H
-#endif
+#ifdef MVLC_GENERATOR
+#include "DeviceCommand.h"
 #endif
 
 // Forward class definitions:
@@ -41,6 +33,12 @@
 class CReadoutModule;
 class CVMUSB;
 class CVMUSBReadoutList;
+
+#ifdef MVLC_GENERATOR
+namespace XXUSB {
+  class CConfigurableObject;
+}
+#endif
 
 /*!
   The C3804 class represents the SIS C3804 scaler module.  This module will be readout
@@ -75,24 +73,36 @@ Option       Value type       Default    Meaning
 class C3804 : public CReadoutHardware
 {
 private:
+#ifdef MVLC_GENERATOR
+  XXUSB::CConfigurableObject* m_pConfiguration;
+#else
   CReadoutModule*   m_pConfiguration;
+#endif
 public:
   C3804();
-  C3804(const C3804& rhs);
   virtual ~C3804();
+#ifdef MVLC_GENMERATOR
+private:
+#endif  
+  C3804(const C3804& rhs);
   C3804& operator=(const C3804& rhs);
-
+public:
 private:
   int operator==(const C3804& rhs) const;
   int operator!=(const C3804& rhs) const;
 
 
 public:
+#ifdef MVLC_GENERATOR
+  virtual void onAttach(XXUSB::CConfigurableObject& configuration);
+#else
   virtual void onAttach(CReadoutModule& configuration);
+#endif
   virtual void Initialize(CVMUSB& controller);
   virtual void addReadoutList(CVMUSBReadoutList& list);
+#ifndef MVLC_GENERATOR
   virtual CReadoutHardware* clone() const;
-
+#endif
 private:
   uint32_t   getIntegerParameter(std::string name) const;
   bool       getBoolParameter(std::string name)    const;
@@ -103,4 +113,20 @@ private:
   void       checkWrite(CVMUSB& controller, uint32_t address, uint32_t value);
 };   
 
+#ifdef MVLC_GENERATOR
+
+/**
+ *  @class SIS3804Command 
+ *   DeviceCommand Deriviation to creat modules wrapping C3804 drivers
+ */
+
+ class SIS3804Command : public DeviceCommand {
+public:
+  SIS3804Command(CTCLInterpreter& interp, TCLConfigParser& parser);
+  virtual ~SIS3804Command();
+
+protected:
+  CReadoutModule* createDevice(std::string name);
+ };
+#endif
 #endif
